+++ src/OneBot/Util/Utils.php
@@ -2,10 +2,8 @@
 
 declare(strict_types=1);
 
-namespace OneBot\V12;
+namespace OneBot\Util;
 
-use OneBot\V12\Action\ActionBase;
-use OneBot\V12\Exception\OneBotFailureException;
 use PASVL\Validation\ValidatorBuilder;
 
 class Utils
@@ -41,35 +39,6 @@ class Utils
             $result = $message;
         }
         return $result;
-    }
-
-    /**
-     * @throws OneBotFailureException
-     */
-    public static function getActionFuncName(ActionBase $handler, string $action)
-    {
-        if (isset(ActionBase::$core_cache[$action])) {
-            return ActionBase::$core_cache[$action];
-        }
-        if (isset(ActionBase::$ext_cache[$action])) {
-            return ActionBase::$ext_cache[$action];
-        }
-        if (substr(
-            $action,
-            0,
-            strlen(OneBot::getInstance()->getPlatform()) + 1
-        ) === (OneBot::getInstance()->getPlatform() . '.')) {
-            $func = self::separatorToCamel('ext_' . substr($action, strlen(OneBot::getInstance()->getPlatform()) + 1));
-            if (method_exists($handler, $func)) {
-                return ActionBase::$ext_cache[$action] = $func;
-            }
-        } else {
-            $func = self::separatorToCamel('on_' . $action);
-            if (method_exists($handler, $func)) {
-                return ActionBase::$core_cache[$action] = $func;
-            }
-        }
-        throw new OneBotFailureException(RetCode::UNSUPPORTED_ACTION);
     }
 
     /**
