+++ src/OneBot/V12/Action/ReplAction.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace OneBot\V12\Action;
 
+use OneBot\Util\Utils;
 use OneBot\V12\Object\ActionObject;
-use OneBot\V12\Utils;
 
 /**
  * Demo REPL Action Handler.
